Move validators into separate folder
